move the super builder to its own module
